fix webgpu

+++ test/testextra/test_blake3.py
@@ -1,13 +1,13 @@
 import unittest
 from extra.utilities.blake3 import BLAKE3
-from tinygrad.tensor import Tensor
+from tinygrad.tensor import Tensor, dtypes
 
 class TestBLAKE3(unittest.TestCase):
   """Test against official tests from: https://github.com/BLAKE3-team/BLAKE3"""
 
   def _test(self, input_len: int, expected_hash: str):
     input_data = bytes(i % 251 for i in range(input_len))
-    actual = BLAKE3().hash(Tensor(input_data), padded_input_size=1024**2)
+    actual = BLAKE3().hash(Tensor(input_data, dtype=dtypes.uint8), padded_input_size=1024**2)
     self.assertEqual(actual, expected_hash)
 
   def test_empty_input(self):
